+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- Throw IpmiException when errors occur instead of returning false
+- No longer return objects from functions which are not fetching data
+
 ## 1.2.0
 
 - Add getPowerState() method